+++ explorer/src/components/inspector/AddressTableLookupsCard.tsx
