+++ explorer/src/components/inspector/InstructionsSection.tsx
