+++ explorer/src/components/inspector/SignaturesCard.tsx
