+++ explorer/src/components/inspector/SimulatorCard.tsx
